+++ Magic/CHANGELOG.md
@@ -2,12 +2,13 @@
 
 # 7.8.1
 
- - Fix Lockpick spel
+ - Fix Lockpick spell
  - Add invert_distance parameter to Damage action, to allow spells to do less damage up close.
  - Add refund_on_no_target spell property, to refund costs on miss
  - Fix broomsticks in 1.15
  - Fix broomsticks insta-crashing if you're not looking up when mounting
  - Disallow using broomsticks in creative mode (PLEASE STOP REPORTING THIS, THE PLUGIN DOES NOT SUPPORT CREATIVE MODE!)
+ - Fix custom model data not carrying over from an item to a wand icon
 
 # 7.8